backend/pyproject.toml
[project]
name = "pmoves-dox-backend"
version = "0.1.0"
description = "PMOVES Document Intelligence Backend"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "docling>=2.50.0,<3.0",
    "pandas>=2.2.0",
    "openpyxl>=3.1.5,<4.0.0",
    "sqlalchemy>=2.0.25",
    "pydantic-settings>=2.3.0,<3.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "beautifulsoup4>=4.12.2",
    "Pillow>=10.4.0",
    "transformers>=4.41.0,<4.50.0",
    "langextract>=0.1.0",
    "python-docx>=1.1.2,<2.0.0",
    "sentence-transformers>=3.0.1",
    "scikit-learn>=1.4.0",
    "PyYAML>=6.0.1",
    "matplotlib>=3.8.0",
    "watchgod>=0.8.2",
    "sqlmodel>=0.0.16",
    "alembic>=1.13.1",
    "faiss-cpu>=1.8.0",
    "supabase>=2.4.0,<3.0.0",
    "torch>=2.3.0",
    "torchvision>=0.18.0",
    "accelerate>=0.27.0",
    "nats-py>=2.8.0",
    "google-generativeai>=0.4.0",
    "google-cloud-aiplatform>=1.38.0",
    "bitsandbytes>=0.43.0; sys_platform != 'win32'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0,<8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0,<8.0.0",
    "ipykernel>=6.29.0",
    "jupyter>=1.0.0",
]

